WIXBUG 4791:  Fix race in lit.exe/light.exe related to wixlibs and large binaries.

+++ src/tools/wix/BinderFileManager.cs
@@ -22,6 +22,7 @@ namespace Microsoft.Tools.WindowsInstallerXml
     using System.Globalization;
     using System.IO;
     using System.Runtime.InteropServices;
+    using System.Threading;
 
     /// <summary>
     /// Options for building the cabinet.
@@ -270,26 +271,28 @@ namespace Microsoft.Tools.WindowsInstallerXml
                 return false;
             }
 
-            using (FileStream targetStream = File.OpenRead(targetFile))
+            using (FileStream targetStream = RetryFileAction(targetFile, () => { return targetFileInfo.OpenRead(); }))
             {
-                using (FileStream updatedStream = File.OpenRead(updatedFile))
+                using (FileStream updatedStream = RetryFileAction(updatedFile, () => { return updatedFileInfo.OpenRead(); }))
                 {
-                    if (targetStream.Length != updatedStream.Length)
+                    if (RetryFileAction(targetFile, () => { return targetStream.Length; }) != RetryFileAction(updatedFile, () => { return updatedStream.Length; }))
                     {
                         return false;
                     }
 
                     // Using a larger buffer than the default buffer of 4 * 1024 used by FileStream.ReadByte improves performance.
                     // The buffer size is based on user feedback. Based on performance results, a better buffer size may be determined.
-                    byte[] targetBuffer = new byte[16 * 1024];
-                    byte[] updatedBuffer = new byte[16 * 1024];
+                    const int bufferSize = 64 * 1024;
+                    byte[] targetBuffer = new byte[bufferSize];
+                    byte[] updatedBuffer = new byte[bufferSize];
                     int targetReadLength;
                     int updatedReadLength;
+
                     do
                     {
-                        targetReadLength = targetStream.Read(targetBuffer, 0, targetBuffer.Length);
-                        updatedReadLength = updatedStream.Read(updatedBuffer, 0, updatedBuffer.Length);
-                        
+                        targetReadLength = RetryFileAction(targetFile, () => { return targetStream.Read(targetBuffer, 0, targetBuffer.Length); });
+                        updatedReadLength = RetryFileAction(updatedFile, () => { return updatedStream.Read(updatedBuffer, 0, updatedBuffer.Length); });
+
                         if (targetReadLength != updatedReadLength)
                         {
                             return false;
@@ -297,7 +300,7 @@ namespace Microsoft.Tools.WindowsInstallerXml
 
                         for (int i = 0; i < targetReadLength; ++i)
                         {
-                            if (targetBuffer[i] != updatedBuffer[i])
+                            if (0 != memcmp(targetBuffer, updatedBuffer, targetReadLength))
                             {
                                 return false;
                             }
@@ -629,9 +632,20 @@ namespace Microsoft.Tools.WindowsInstallerXml
         /// <param name="overwrite">true if the destination file can be overwritten; otherwise, false.</param>
         public virtual void CopyFile(string source, string destination, bool overwrite)
         {
-            if (overwrite)
+            if (!File.Exists(source))
             {
-                File.Delete(destination);
+                throw new WixFileNotFoundException(source);
+            }
+
+            if (overwrite && File.Exists(destination))
+            {
+                RetryFileAction(destination, () =>
+                {
+                    FileAttributes attributes = File.GetAttributes(destination);
+                    File.SetAttributes(destination, attributes & ~FileAttributes.ReadOnly);
+                    File.Delete(destination);
+                    return true;
+                });
             }
 
             if (!CreateHardLink(destination, source, IntPtr.Zero))
@@ -640,7 +654,13 @@ namespace Microsoft.Tools.WindowsInstallerXml
                 int er = Marshal.GetLastWin32Error();
 #endif
 
-                File.Copy(source, destination, overwrite);
+                // this.core.OnMessage(WixVerboses.CreateDirectory(directory));
+                RetryFileAction(destination, () =>
+                {
+                    Directory.CreateDirectory(Path.GetDirectoryName(destination));
+                    File.Copy(source, destination, overwrite);
+                    return true;
+                });
             }
         }
 
@@ -651,7 +671,29 @@ namespace Microsoft.Tools.WindowsInstallerXml
         /// <param name="destination">The destination file.</param>
         public virtual void MoveFile(string source, string destination)
         {
-            File.Move(source, destination);
+            if (!File.Exists(source))
+            {
+                throw new WixFileNotFoundException(source);
+            }
+
+            if (File.Exists(destination))
+            {
+                RetryFileAction(destination, () =>
+                {
+                    FileAttributes attributes = File.GetAttributes(destination);
+                    File.SetAttributes(destination, attributes & ~FileAttributes.ReadOnly);
+                    File.Delete(destination);
+                    return true;
+                });
+            }
+
+            // this.core.OnMessage(WixVerboses.CreateDirectory(directory));
+            RetryFileAction(destination, () =>
+            {
+                Directory.CreateDirectory(Path.GetDirectoryName(destination));
+                File.Move(source, destination);
+                return true;
+            });
         }
 
         /// <summary>
@@ -735,7 +777,60 @@ namespace Microsoft.Tools.WindowsInstallerXml
             }
         }
 
+        /// <summary>
+        /// Private method to retry a file action that may block because
+        /// another process is working the file.  Retries on
+        /// an <see cref="IOException" />.
+        /// </summary>
+        /// <param name="path">File path to watch.</param>
+        /// <typeparam name="T">Return type of the file action.</typeparam>
+        /// <param name="func">File <c>I/O</c> Delegate to retry.</param>
+        /// <returns>
+        /// Returns the result of the delegate on success.
+        /// </returns>
+        private T RetryFileAction<T>(string path, Func<T> func)
+        {
+            var autoResetEvent = new AutoResetEvent(false);
+            int retries = 3;
+
+            while (true)
+            {
+                try
+                {
+                    return func();
+                }
+                catch (IOException)
+                {
+                    var fsw = new FileSystemWatcher(Path.GetDirectoryName(path)) { EnableRaisingEvents = true };
+
+                    fsw.Changed += (o, e) =>
+                        {
+                            if (Path.GetFullPath(e.FullPath).Equals(Path.GetFullPath(path), StringComparison.OrdinalIgnoreCase))
+                            {
+                                autoResetEvent.Set();
+                            }
+                        };
+
+                    autoResetEvent.WaitOne();
+                }
+                catch (UnauthorizedAccessException)  // for MoveFile and CopyFile
+                {
+                    if (1 > retries)
+                    {
+                        // done trying, permissions/access issues remain
+                        throw;
+                    }
+
+                    // unlikely to improve past 3 retries
+                    --retries;
+                }
+            }
+        }
+
         [DllImport("Kernel32.dll", CharSet = CharSet.Unicode, SetLastError = true)]
         private static extern bool CreateHardLink(string lpFileName, string lpExistingFileName, IntPtr lpSecurityAttributes);
+
+        [DllImport("msvcrt.dll", CallingConvention = CallingConvention.Cdecl)]
+        static extern int memcmp(byte[] buf1, byte[] buf2, long count);
     }
 }
